fix: credit note

+++ src/main/java/dev/arias/huapaya/repair_shop/presentation/dto/credit_note/CreditNoteFindOneDTO.java
@@ -34,7 +34,7 @@ public class CreditNoteFindOneDTO {
 
     private String serie;
 
-    private Integer number;
+    private Long number;
 
     private BigDecimal taxAmount;
